Fix #4: Make sly-stepper contrib a no-op for ABCL.
Gray streams on ABCL seem to be slightly off and do not work well with
the slynk-gray.  This is an old problem in both SLY and SLIME.

Until we figure this out, let's make sly-stepper a no-op for ABCL so
that the contrib can remained globally enabled (for other
implementations).

Co-authored-by: João Távora <joaotavora@gmail.com>